+++ src/test/java/racingcar/CarTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.Car;
 import racingcar.movestrategy.MoveStrategy;
 
 public class CarTest {
