+++ src/test/java/racingcar/CarsTest.java
@@ -8,6 +8,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 import racingcar.movestrategy.MoveStrategy;
 
 public class CarsTest {
